hoành thành bài tập lớn
 hoàn thành bài tập lớn a-z

+++ Book-v2/src/admin_controller/DeleteCategory.java
@@ -2,6 +2,7 @@ package admin_controller;
 
 import java.io.IOException;
 
+import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -9,6 +10,7 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import dao.CategoryDAO;
+import model.User;
 
 public class DeleteCategory extends HttpServlet {
 	private static final long serialVersionUID = 1L;
@@ -28,11 +30,20 @@ public class DeleteCategory extends HttpServlet {
 			response.sendRedirect("/Book/admin/Login");
 		} else {
 
-			Long id = Long.parseLong(request.getParameter("id"));
+			String roleAdmin = (String) session.getAttribute("role");
 
-			new CategoryDAO().deleteById(id);
+			if (roleAdmin.equals("" + User.GIAMDOC) || roleAdmin.equals("" + User.QUANLYTHELOAI)) {
 
-			response.sendRedirect("/Book/admin/listCategory");
+				Long id = Long.parseLong(request.getParameter("id"));
+				new CategoryDAO().deleteById(id);
+				response.sendRedirect("/Book/admin/listCategory");
+
+			} else {
+				
+				RequestDispatcher rd = getServletContext().getRequestDispatcher("/admin/dasboard");
+				rd.forward(request, response);
+			}
+
 		}
 	}
