+++ Book-v2/src/admin_controller/DeleteUser.java
@@ -1,7 +1,9 @@
 package admin_controller;
 
 import java.io.IOException;
+import java.sql.SQLException;
 
+import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -9,6 +11,7 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import dao.UserDAO;
+import model.User;
 
 public class DeleteUser extends HttpServlet {
 	private static final long serialVersionUID = 1L;
@@ -27,19 +30,47 @@ public class DeleteUser extends HttpServlet {
 			// User is not logged in.
 			response.sendRedirect("/Book/admin/Login");
 		} else {
-			Long id = Long.parseLong(request.getParameter("id"));
-			
-			String type = request.getParameter("type");
-			
-			
-			new UserDAO().deleteById(id);
-			if(type.equals("customer")) {
-				response.sendRedirect("/Book/admin/listCustomer");
+
+			String roleAdmin = (String) session.getAttribute("role");
+
+			if (roleAdmin.equals("" + User.GIAMDOC) || roleAdmin.equals("" + User.QUANLYNHANVIEN)) {
+
+				Long id = Long.parseLong(request.getParameter("id"));
+				String type = (request.getParameter("type") == null) ? "" : request.getParameter("type");
+				
+				// delete khách hàng.
+				if(type.equals("customer")) {
+					System.out.println(id);
+					try {
+						new UserDAO().deleteById(id);
+						response.sendRedirect("/Book/admin/listCustomer");
+					} catch (SQLException e) {
+						response.sendRedirect("/Book/admin/listCustomer?error=delete");
+					}
+					
+				} else {
+
+					Long idLogin = Long.parseLong(session.getAttribute("idAdmin") + "");
+
+					User u = new UserDAO().getUserById(id);
+					
+					if (u.getCreateBy().equals(idLogin)) {
+						try {
+							new UserDAO().deleteById(id);
+							response.sendRedirect("/Book/admin/listUser");
+						} catch(Exception e) {
+							response.sendRedirect("/Book/admin/listUser?error=delete");
+						}
+					} else {
+						response.sendRedirect("/Book/admin/listUser?error=delete");
+					}
+				}
+
 			} else {
-				response.sendRedirect("/Book/admin/listUser");
+				RequestDispatcher rd = getServletContext().getRequestDispatcher("/admin/dasboard");
+				rd.forward(request, response);
 			}
 
-			
 		}
 	}
